move process_many to tests/common
