+++ core-processor/src/lib.rs
@@ -52,6 +52,6 @@ pub use ext::Ext;
 pub use ext::ProcessorExt;
 pub use handler::handle_journal;
 pub use id::{next_message_id, next_system_reply_message_id, BlakeMessageIdGenerator};
-pub use processor::{process, process_many};
+pub use processor::process;
 
 use gear_core::message::ExitCode;
